requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
torch
faiss-cpu
pickle5
requests
langchain
huggingface_hub